refs #466 merged [2152] from branches/0.8
git-svn-id: https://devel.codership.com/svn/galera/trunk@2153 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galera/src/trx_handle.hpp
@@ -305,7 +305,7 @@ namespace galera
         bool                   committed_;
         long                   gcs_handle_;
         void*                  action_;
-        long long              timestamp_;
+        int64_t                timestamp_;
         Mac                    mac_;
 
         //
